+++ src/app/master-data/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation";
 
 export default function MasterDataPage() {
-  redirect("/master-data/locations");
+  redirect("/master-data/products");
 }